renaming index.js into home.js
